+++ src/overworld/interaction.rs
@@ -1,6 +1,7 @@
 use bevy::prelude::*;
 
 use crate::core::GameState;
+use crate::ui::inventory_window::InventoryWindowOpen;
 use crate::exploration::{
     set_exploration_prompt, set_exploration_zone_label, EXPLORATION_PROMPT_MOVE_INTERACT,
 };
@@ -12,6 +13,7 @@ use super::setup::{OverworldPromptLabel, OverworldZoneLabel};
 
 pub fn overworld_interaction(
     keyboard: Res<ButtonInput<KeyCode>>,
+    inventory: Res<InventoryWindowOpen>,
     layout: Res<OverworldLayout>,
     cooldown: Res<MapTransitionCooldown>,
     player: Query<(&Transform, &OverworldVelocity), With<OverworldPlayer>>,
@@ -22,6 +24,10 @@ pub fn overworld_interaction(
     let Ok((transform, velocity)) = player.get_single() else {
         return;
     };
+
+    if inventory.0 {
+        return;
+    }
 
     let position = transform.translation.truncate();
     let zone = layout.zone_at(position);
